Changed representations of words to matrices.
This included a change of feature values from strings to integers.
Also changed Gen so that changes record only these three features of the
segments that were affected: vocalic, consonantal, and sonorant.
The matrix change required changes to Markedness.pick_unique_pattern(), which
needed work anyway. The current implementation is better in theory, but it uses
count(), which doesn't work on matrices, so that needs to be translated to
something that works.