Added save option for Input, added files, debugged.
The input is read from a file if the right one exists, or else made and then
written to a file.

Mapping and FeatureDict are now in their own files so that
cPickle will work.

Debugged and sped up change_feature_value, ungrammaticalize, MarkednessAligned.
Removed make_tableaux. It's now implemented in make_input, which sped it up.