Fixed constraint duplication, added faithful cand.
Duplicate faithfulness constraints were allowed, because the __eq__ method
wasn't working. The test for duplicates is now based on the constraint
attribute, which works.

The report was also writing all of the constraints under "constraints added."
Now it will only write the new constraints there.

Gen will now make sure that the faithful candidate is generated if it's not the
grammatical mapping.