Added stem parameter, refactored reporting.
Removed changing of induction_freq to 0 on odd iterations.
Reporting is a little easier to follow, but I added a tracking of error numbers
for graphing that I haven't made use of yet.
May not need to because I'm using a smaller training set now, 1/10th of the
data. This gives a finer grained learning curve.